backend-examples/python-flask/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7